Move ValidationActionRenderer to Libplanet project
so that other projects than Libplanet.Tests can this too.
