+++ Libplanet.Tests/TestUtils.cs
@@ -11,11 +11,11 @@ using Libplanet.Assets;
 using Libplanet.Blockchain;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Blockchain.Renderers;
+using Libplanet.Blockchain.Renderers.Debug;
 using Libplanet.Blocks;
 using Libplanet.Crypto;
 using Libplanet.Store;
 using Libplanet.Store.Trie;
-using Libplanet.Tests.Common;
 using Libplanet.Tx;
 using Xunit;
 using static Libplanet.Blockchain.KeyConverters;
